host/event_manage_proj/requirement.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.3.2
cmake==3.30.3
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
Django==5.1.1
django-phonenumber-field==8.0.0
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.14.0
fonttools==4.53.1
idna==3.10
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.11.0
kiwisolver==1.4.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==24.0
pandas==2.2.2
parso==0.8.4
phonenumbers==8.13.47
pillow==10.4.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pyzmq==26.0.3
razorpay==1.4.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.3
virtualenv==20.26.2
wcwidth==0.2.13
